feat: replace gpm dep with termion
TODO: migrate from slang to some similar rust lib too

+++ src/aaimgheight.rs
@@ -1,7 +1,7 @@
 use super::aastructs::*;
 
-pub fn aa_imgheight(a: *mut aa_context) -> i64 {
+pub fn aa_imgheight(a: *mut aa_context) -> i32 {
     unsafe {
-        return (*a).imgheight;
+        return (*a).imgheight as i32;
     }
 }
