+++ src/aaimgwidth.rs
@@ -1,7 +1,7 @@
 use super::aastructs::*;
 
-pub fn aa_imgwidth(a: *mut aa_context) -> i64 {
+pub fn aa_imgwidth(a: *mut aa_context) -> i32 {
     unsafe {
-        return (*a).imgwidth;
+        return (*a).imgwidth as i32;
     }
 }
